# api/openapi.json
{
  "components": {
    "schemas": {
      "AddDraftAttachmentInput": {
        "type": "object",
        "properties": {
          "draft_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "data_base64": {
            "type": "string"
          },
          "inline": {
            "type": "boolean"
          }
        },
        "title": "AddDraftAttachmentInput",
        "required": [
          "name",
          "data_base64"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": true
          },
          "trace_id": {
            "type": "string"
          }
        },
        "title": "ErrorResponse",
        "required": [
          "code",
          "message",
          "trace_id"
        ]
      },
...
      "RenderedMail": {
        "type": "object",
        "properties": {
          "body_html": {
            "type": "string"
          },
          "body_text": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "template": {
            "type": "string"
          },
          "warnings": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "title": "RenderedMail",
        "required": [
          "body_text",
          "format",
          "template"
        ],
        "additionalProperties": false
      },
...
      "SettingsPatch": {
        "type": "object",
        "properties": {
          "values": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "secrets": {
            "type": [
              "null",
              "object"
            ],
            "description": "Write-only replacement secrets. Never returned by settings reads.",
            "writeOnly": true,
            "additionalProperties": {
              "type": "string"
            }
          },
          "locks": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "boolean"
            }
          },
          "reset": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "revision": {
            "type": "string"
          }
        },
        "title": "SettingsPatch"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Scoped Postra MCP key or supported deployment bearer credentials. Capabilities and ownership are enforced by the server.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Browser session; mutations additionally require same-origin CSRF checks.",
        "in": "cookie",
        "name": "postra_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Generated from Go DTOs for settings, Ask, rendering, drafts, preview, attachment mutation and signatures only. This is not the complete Postra REST/MCP API. /api/v1 is canonical; /api is a compatibility alias using the same handlers. Sending/approval, binary downloads, auth, admin probes, and other operations are intentionally outside this snapshot. JSON schemas describe wire shapes, not every dynamic validation or permission rule.",
    "title": "Postra canonical core API — partial coverage",
    "version": "1"
  },
...
  "openapi": "3.1.0",
  "paths": {
    "/accounts/{id}/preferences": {
      "get": {
        "operationId": "accountPreferencesGet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Effective preferences for an owned account"
      },
      "patch": {
        "operationId": "accountPreferencesPatch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Update an owned account's unlocked preferences"
      }
    },
    "/admin/configuration": {
      "get": {
        "operationId": "adminConfigurationGet",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Administrator catalog with effective values and lock/source metadata"
      },
      "patch": {
        "operationId": "adminConfigurationPatch",
        "parameters": [
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Administrator settings and write-only secret update"
      }
    },
    "/drafts": {
      "post": {
        "operationId": "draftCreate",
        "parameters": [
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Create a draft without approving or sending"
      }
    },
    "/drafts/{id}": {
      "get": {
        "operationId": "draftGet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Read the current version of an owned draft"
      },
      "patch": {
        "operationId": "draftUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Create a new immutable version; invalidates earlier approval"
      }
    },
    "/drafts/{id}/attachments": {
      "post": {
        "operationId": "draftAttachmentAdd",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Multipart upload only: true inserts a validated PNG/JPEG/GIF as a CID image. JSON uploads use their inline field.",
            "in": "query",
            "name": "inline",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDraftAttachmentInput"
              }
            },
            "multipart/form-data": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Add a scanned attachment in a new draft version"
      }
    },
    "/drafts/{id}/attachments/{attachment}": {
      "delete": {
        "operationId": "draftAttachmentRemove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attachment",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Remove an attachment from a new version, retaining history"
      }
    },
    "/drafts/{id}/preview": {
      "post": {
        "operationId": "draftSendPreview",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendPreview"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Review canonical send payload and policies; does not issue approval"
      }
    },
    "/mail/render": {
      "post": {
        "operationId": "mailRender",
        "parameters": [
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderMailInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderedMail"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Render locally; optional explicit smart_format uses AI but never sends"
      }
    },
    "/mail/templates": {
      "get": {
        "operationId": "mailTemplates",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MailTemplate"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "List deterministic built-in templates"
      }
    },
    "/mcp/connection": {
      "get": {
        "operationId": "mcpConnectionGet",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPOAuthConnectionInfo"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Authenticated user's public MCP connection instructions; no credentials or health assertion"
      }
    },
    "/preferences": {
      "get": {
        "operationId": "preferencesGet",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Current user's effective preferences"
      },
      "patch": {
        "operationId": "preferencesPatch",
        "parameters": [
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Update current user's unlocked preferences"
      }
    },
    "/qa": {
      "post": {
        "operationId": "mailAsk",
        "parameters": [
          {
            "description": "Required with cookie authentication, together with an allowed same-origin Origin header. Bearer requests follow their separate scope policy.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResult"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Classified error; dynamic ownership, scopes, policy and revision checks remain authoritative. Legacy error alias may also be present."
          }
        },
        "summary": "Answer a question using bounded owned-mail/work/action evidence"
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Canonical API v1",
      "url": "/api/v1"
    }
  ],
...
}